+++ uninstall.php
@@ -39,6 +39,7 @@ delete_option('ihumbak_wrs_email_body');
 delete_option('ihumbak_wrs_email_from_name');
 delete_option('ihumbak_wrs_email_from_email');
 delete_option('ihumbak_wrs_email_reply_to');
+delete_option('ihumbak_wrs_email_coupon_id');
 delete_option('ihumbak_wrs_email_followups');
 
 // Delete transients